refactor: fare 패키지 분리
