+++ src/main/java/nextstep/subway/fare/domain/FarePolicy.java
@@ -1,4 +1,4 @@
-package nextstep.subway.path.domain;
+package nextstep.subway.fare.domain;
 
 public interface FarePolicy {
