+++ src/main/java/nextstep/subway/fare/domain/LineSurchargePolicy.java
@@ -1,4 +1,4 @@
-package nextstep.subway.path.domain;
+package nextstep.subway.fare.domain;
 
 import nextstep.subway.line.domain.Line;
