+++ src/main/java/nextstep/subway/path/application/PathService.java
@@ -3,8 +3,8 @@ package nextstep.subway.path.application;
 import nextstep.subway.auth.domain.LoginMember;
 import nextstep.subway.line.domain.LineRepository;
 import nextstep.subway.line.domain.Lines;
-import nextstep.subway.path.domain.FarePolicy;
-import nextstep.subway.path.domain.FarePolicyFactory;
+import nextstep.subway.fare.domain.FarePolicy;
+import nextstep.subway.fare.domain.FarePolicyFactory;
 import nextstep.subway.path.domain.PathFinder;
 import nextstep.subway.path.domain.ShortestPath;
 import nextstep.subway.path.dto.PathResponse;
