+++ src/main/java/nextstep/subway/path/dto/PathResponse.java
@@ -1,6 +1,6 @@
 package nextstep.subway.path.dto;
 
-import nextstep.subway.path.domain.Fare;
+import nextstep.subway.fare.domain.Fare;
 import nextstep.subway.path.domain.ShortestPath;
 import nextstep.subway.station.domain.Station;
 import nextstep.subway.station.dto.StationResponse;
