+++ src/test/java/nextstep/subway/fare/domain/DistanceFarePolicyTest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.path.domain;
+package nextstep.subway.fare.domain;
 
 import nextstep.subway.line.domain.Distance;
 import org.junit.jupiter.params.ParameterizedTest;
